restructures components folder
