+++ src/client/src/views/USPage.vue
@@ -1,11 +1,11 @@
 <template>
   <div class="USPage fadeIn">
-    <USBody firstInputName="County" secondInputName="State" apiEndPoint="/USData" />
+    <Body firstInputName="County" secondInputName="State" apiEndPoint="/USData" />
   </div>
 </template>
 
 <script>
-import USBody from "../components/US_World_Body";
+import Body from "../components/US_World/Body";
 
 export default {
   name: "USPage",
@@ -14,7 +14,7 @@ export default {
   },
   methods: {},
   components: {
-    USBody
+    Body
   }
 };
 </script>
